Add unsafe item service

+++ src/main/java/io/github/jaswinder/distributedlockdemo/repository/ItemAvailabilityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItemAvailabilityRepository extends JpaRepository<ItemAvailability, Long> {
+public interface ItemAvailabilityRepository extends JpaRepository<ItemAvailability, Integer> {
 }
